backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
django-cloudinary-storage>=0.3
cloudinary>=1.40
pusher>=3.3
stripe>=8.0
django-filter>=24.0
python-decouple>=3.8
gunicorn>=22.0
django-fsm>=3.0
pillow>=10.0
